Divided the generator/content.js file in two (content.js & block.js). They call each other, but there is no circular dependency
